Simplify concourse auth check.
As of concourse 4.0, the list teams endpoint only shows teams that the
current user belongs to. This means we can't use unauthenticated
requests to describe teams and authentication. Instead, this patch
verifies that unauthenticated requests are rejected by sending an
unauthenticated request and asserting that the status code is a 401.